+++ R/read_out.R
@@ -91,7 +91,7 @@ please leave patient_cut_data empty, in which case a default value of NULL will 
 
     for (name in names(patient_cut_data)) {
       if (is.data.frame(patient_cut_data[[name]]) && nrow(patient_cut_data[[name]]) == 0) {
-        print(paste("The dataset", name, "is empty"))
+        print(paste("The", name, "dataset is empty, no cut will be performed"))
       }
     }
 
@@ -118,7 +118,7 @@ leave date_cut_data empty, in which case a default value of NULL will be used."
     # Check for empty data frames
     for (name in names(date_cut_data)) {
       if (is.data.frame(date_cut_data[[name]]) && nrow(date_cut_data[[name]]) == 0) {
-        print(paste("The dataset", name, "is empty"))
+        print(paste("The", name, "dataset is empty, no cut will be performed"))
       }
     }
 
@@ -136,6 +136,11 @@ leave date_cut_data empty, in which case a default value of NULL will be used."
     }
   }
   if (!is.null(dm_cut)) {
+    if (nrow(dm_cut) == 0L) {
+      print("The dm dataset is empty")
+      dm_cut$DCUT_TEMP_REMOVE <- character()
+      dm_cut$DCUT_TEMP_DTHCHANGE <- character()
+    }
     assert_data_frame(dm_cut,
       required_vars = exprs(USUBJID, DCUT_TEMP_REMOVE, DCUT_TEMP_DTHCHANGE)
     )
@@ -147,6 +152,17 @@ Note: If you have not used or do not with to view the SDTMv domains where no cut
 applied, then please leave no_cut_list empty, in which case a default value of NULL will be
 used."
     )
+
+    # Check for empty data frames
+    for (name in names(no_cut_list)) {
+      if (is.data.frame(no_cut_list[[name]]) && nrow(no_cut_list[[name]]) == 0) {
+        print(paste("The", name, "dataset is empty"))
+      }
+    }
+
+    # Don't include empty datasets in report
+    no_cut_list <- no_cut_list[which(lapply(no_cut_list, nrow) != 0)]
+
     for (i in seq_along(no_cut_list)) {
       assert_data_frame(no_cut_list[[i]])